Numerous fixes made to install script
Various bugs were solved centering around the user's .source file.
The script was not able to give the current script user session the location of the go bin directory or go home directory.
Script now has a section that checks the OpenGL version to ensure it is at least the minimum compatible v3.3